# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MultiTrack Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Admin dashboard metrics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, all]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMetrics"
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: Admin list all products
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, draft, featured]
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
  /admin/products/{id}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: Admin list all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, cancelled, refunded, chargeback]
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin list users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [visitor, buyer, member, premium, admin, support]
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Admin get user details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user status or role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSummary"
...
  /admin/rateios/{id}/status:
    patch:
      operationId: adminUpdateRateioStatus
      tags: [admin]
      summary: Admin update rateio status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRateioStatusBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rateio"
  /admin/coupons:
    get:
      operationId: adminListCoupons
      tags: [admin]
      summary: Admin list coupons
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: adminCreateCoupon
      tags: [admin]
      summary: Create a coupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /admin/coupons/{id}:
    patch:
      operationId: adminUpdateCoupon
      tags: [admin]
      summary: Update a coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /admin/credits/adjust:
    post:
      operationId: adminAdjustCredits
      tags: [admin]
      summary: Admin credit adjustment for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreditAdjustBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/downloads/{grantId}/reissue:
    post:
      operationId: adminReissueDownload
      tags: [admin]
      summary: Admin reissue download link for a grant
      parameters:
        - name: grantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadLink"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        isBlocked:
          type: boolean
        ordersCount:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, isBlocked, ordersCount, totalSpent, createdAt]
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/UserSummary"
        - type: object
          properties:
            subscription:
              $ref: "#/components/schemas/Subscription"
              nullable: true
            creditBalance:
              type: integer
            recentOrders:
              type: array
              items:
                $ref: "#/components/schemas/Order"
            notes:
              type: string
              nullable: true
    AdminUpdateUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [buyer, member, premium, admin, support]
        isBlocked:
          type: boolean
        notes:
          type: string

    # ── Category ──────────────────────────────────────────────────────────
...
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            description:
              type: string
            tracks:
              type: array
              items:
                type: string
            formats:
              type: array
              items:
                type: string
            fileSizeMb:
              type: number
              nullable: true
            previewAudioUrl:
              type: string
              nullable: true
            videoUrl:
              type: string
              nullable: true
            tags:
              type: array
              items:
                type: string
            licenceSummary:
              type: string
              nullable: true
            category:
              $ref: "#/components/schemas/Category"
              nullable: true
            averageRating:
              type: number
              nullable: true
            reviewCount:
              type: integer
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [products, total, page, limit]
    AdminProductBody:
      type: object
      properties:
        name:
          type: string
        artist:
          type: string
        genre:
          type: string
        categoryId:
          type: integer
          nullable: true
        tonality:
          type: string
          nullable: true
        bpm:
          type: integer
          nullable: true
        duration:
          type: string
          nullable: true
        description:
          type: string
        tracks:
          type: array
          items:
            type: string
        formats:
          type: array
          items:
            type: string
        fileSizeMb:
          type: number
          nullable: true
        coverUrl:
          type: string
          nullable: true
        previewAudioUrl:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        price:
          type: number
        promoPrice:
          type: number
          nullable: true
        promoPriceStartAt:
          type: string
          format: date-time
          nullable: true
        promoPriceEndAt:
          type: string
          format: date-time
          nullable: true
        quality:
          type: string
          enum: [premium, standard, backing_track, demo]
        status:
          type: string
          enum: [active, inactive, draft, featured]
        isFeatured:
          type: boolean
        availableForSale:
          type: boolean
        availableForSubscription:
          type: boolean
        creditsRequired:
          type: integer
        licenceSummary:
          type: string
          nullable: true
      required: [name, artist, genre, description, price, quality, status]

    # ── Cart ──────────────────────────────────────────────────────────────
...
    Order:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, paid, cancelled, refunded, chargeback]
        paymentMethod:
          type: string
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        couponCode:
          type: string
          nullable: true
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, status, paymentMethod, subtotal, discount, total, itemCount, createdAt]
...
    DownloadLink:
      type: object
      properties:
        url:
          type: string
        expiresAt:
          type: string
          format: date-time
      required: [url, expiresAt]

    # ── Subscriptions ─────────────────────────────────────────────────────
...
    Rateio:
      type: object
      properties:
        id:
          type: integer
        songName:
          type: string
        artist:
          type: string
        genre:
          type: string
          nullable: true
        status:
          type: string
          enum: [suggested, in_quotation, open, goal_reached, awaiting_payment, payment_confirmed, in_progress, completed, cancelled, refunded]
        targetAmount:
          type: number
          nullable: true
        amountPerParticipant:
          type: number
          nullable: true
        minParticipants:
          type: integer
          nullable: true
        currentParticipants:
          type: integer
        minPercentage:
          type: number
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        coverUrl:
          type: string
          nullable: true
        isParticipating:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, songName, artist, status, currentParticipants, isParticipating, createdAt]
...
    AdminRateioStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [in_quotation, open, goal_reached, awaiting_payment, payment_confirmed, in_progress, completed, cancelled, refunded]
        targetAmount:
          type: number
          nullable: true
        amountPerParticipant:
          type: number
          nullable: true
        minParticipants:
          type: integer
          nullable: true
        minPercentage:
          type: number
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        cancelReason:
          type: string
          nullable: true
      required: [status]

    # ── Wishlist ──────────────────────────────────────────────────────────
...
    DashboardMetrics:
      type: object
      properties:
        totalRevenue:
          type: number
        revenueByChannel:
          type: object
          properties:
            singlePurchase:
              type: number
            subscription:
              type: number
            bundle:
              type: number
            rateio:
              type: number
          required: [singlePurchase, subscription, bundle, rateio]
        mrr:
          type: number
        churnRate:
          type: number
        averageOrderValue:
          type: number
        totalOrders:
          type: integer
        totalUsers:
          type: integer
        premiumUsers:
          type: integer
        totalDownloads:
          type: integer
        topProducts:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              artist:
                type: string
              sales:
                type: integer
              revenue:
                type: number
            required: [id, name, artist, sales, revenue]
        rateioStats:
          type: object
          properties:
            open:
              type: integer
            completed:
              type: integer
            cancelled:
              type: integer
          required: [open, completed, cancelled]
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
      required: [totalRevenue, revenueByChannel, mrr, churnRate, averageOrderValue, totalOrders, totalUsers, premiumUsers, totalDownloads, topProducts, rateioStats, recentOrders]